// src/app/main/models/organisation.model.ts
export interface Organisation {
  action: string;
  country: string;
  fullLegalName: string;
  shortName: string;
  status: string;
  industry: string;
  pricingSource: string;
  location: string;
  orgId: string;
  countryOfLocation: string;
  privateKey: string;
  publicKey: string;
  ticker: string;
}